backend/requirements.txt
Flask
requests
flask_cors
gunicorn
uvicorn
psycopg2-binary
python-dotenv